Safari-609.1.11.1/LayoutTests/media/modern-media-controls/pip-support/pip-support-click-expected.txt

Testing the PiPSupport behavior when clicking the PiP button.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaController.controls.pipButton.element.getBoundingClientRect().width became different from 0
PASS media.webkitPresentationMode is "picture-in-picture"
PASS successfullyParsed is true

TEST COMPLETE
